agents-communication-sw/wholesaler-agent/requirements.txt
a2a-sdk>=0.2.6
click>=8.1.8
dotenv>=0.9.9
httpx>=0.28.1
langchain-google-genai>=2.1.4
langgraph>=0.4.1
pydantic>=2.11.4
python-dotenv>=1.1.0
uvicorn>=0.34.2
langchain-core>=0.3.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
ruff>=0.1.0
